requirements.txt
openai==0.27.0
python-dotenv==1.0.0
rich==13.3.2
praw==7.7.0
googlesearch-python==1.1.0
promptlayer==0.1.77
discord==2.2.2
pydub==0.25.1
PyNaCl==1.5.0